extract `ew_substrate_fixed` package
