+++ app/test/service/tx/lib/send_tx_dart_test.dart
@@ -1,9 +1,8 @@
 // ignore_for_file: avoid_print
 
-// @Skip('Skip these tests as they need a specific setup.')
+@Skip('Skip these tests as they need a specific setup.')
 
 import 'dart:async';
-import 'dart:math' as math;
 import 'package:encointer_wallet/service/tx/lib/src/send_tx_dart.dart';
 
 import 'package:ew_polkadart/ew_polkadart.dart';
@@ -56,53 +55,5 @@ void main() {
 
       await completer.future.then((_) => sub.cancel());
     });
-
-    test('get ceremony reward', () async {
-      final provider = Provider.fromUri(Uri.parse('wss://gesell.encointer.org'));
-
-      await Future<void>.delayed(const Duration(milliseconds: 500));
-
-      final reward = await EncointerKusama(provider).query.encointerCeremonies.ceremonyReward();
-
-      final value = reward.bits;
-
-      final parser = fixedPointParser(64, 64);
-      final parsed = parser(value);
-
-      print('reward: $parsed');
-    });
   });
-}
-
-
-
-
-typedef FixedPointParser = double Function(BigInt upper, [int precision]);
-
-FixedPointParser fixedPointParser(int upper, int lower) {
-  final len = upper + lower;
-
-  return (BigInt raw, [int precision = 0]) {
-
-    raw = raw.toSigned(len);
-
-    final bits = raw.toRadixString(2).padLeft(len, '0');
-    final lowerBits = bits.substring(bits.length - lower);
-    final upperBits = bits.substring(0, bits.length - lower);
-
-    print('bits: $bits');
-
-    print('lowerBits: $lowerBits');
-    print('upperBits: $upperBits');
-    final floatPart = lowerBits
-        .split('')
-        .asMap()
-        .entries
-        .map((entry) => entry.value == '1' ? 1 / math.pow(2, entry.key + 1) : 0)
-        .reduce((acc, val) => acc + val);
-
-    final decimalPart = upperBits.isNotEmpty ? int.parse(upperBits, radix: 2) : 0;
-
-    return (decimalPart + (raw.isNegative ? -floatPart : floatPart)).toDouble();
-  };
 }
