+++ packages/ew_substrate_fixed/README.md
@@ -1,0 +1,3 @@
+# Substrate Fixed
+
+Implements substrate fixed support for some fixed point types that are used in encointer, like U64F64,
