+++ packages/ew_substrate_fixed/analysis_options.yaml
@@ -1,0 +1,5 @@
+include: package:very_good_analysis/analysis_options.3.1.0.yaml
+
+linter:
+  rules:
+    lines_longer_than_80_chars: false
