+++ packages/ew_substrate_fixed/lib/substrate_fixed.dart
@@ -1,0 +1,3 @@
+library substrate_fixed;
+
+export 'src/substrate_fixed.dart';
