+++ packages/ew_substrate_fixed/pubspec.yaml
@@ -1,0 +1,11 @@
+name: ew_substrate_fixed
+description: Implements fixed point handling in dart
+version: 0.0.1
+publish_to: none
+
+environment:
+  sdk: ">=3.0.3 <4.0.0"
+
+dev_dependencies:
+  test: ^1.24.3
+  very_good_analysis: ^5.0.0+1
